Use metadata files instead of boxfiles
To better make use of Vagrants built-in support for checksums
and versions, we now use the .json metadata files in Vagrant instead
of adding boxfiles directly.

This change also adds a hostname that is separate from the boxname
to make it quicker to refer to vagrantboxes on the cli.

The Debian-based guests were also renamed so that all have the same
distro-version-arch name-scheme, instead of some having codenames.